requirements.txt
boto3>=1.26.0
psutil>=5.6.7
pytz>=2022.1
python-dotenv>=1.0.0